Fix the Simulator with C++ simulator values
Getting the same values from C++ simulator program using the same
initial condition. The calling of functions to change
stage was wrong and addind time was fixed to be added at the end of the
main loop corresponding to the actual stage. Part of print the chart was
commented for analyzing the changing of background corresponding to the
stage but it's not working.